pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "SGrPythontks4r"
dynamic = ["version"]
description = ""
readme = ""
requires-python = ">=3.9"
license = { text = "BSD" }
authors = [{ name = "Robin Schoch" }]
urls = { "Homepage" = "https://github.com/SmartGridready/SGrPython" }
dependencies = [
    "Jinja2>=3.0.0,<4.0.0",
    "jmespath>=1.0.0,<2.0.0",
    "pymodbus>=3.0.0,<4.0.0",
    "setuptools>=68.0.0,<69.0.0",
    "xsdata>=22.0.0,<23.0.0",
    "aiohttp>=3.0.0,<4.0.0",
    "certifi",
    "cachetools>=5.0.0,<6.0.0",
    "SGrSpecificationPythontks4r==0.1.2",
    "setuptools_scm",
]

[tool.setuptools.packages.find]
include = [
    "sgr_library",
    "sgr_library.api",
    "sgr_library.validators",
    "sgr_library.converters",
]

[tool.ruff]
